backend/requirements.txt
fastapi
uvicorn[standard]
alembic
pydantic
python-jose[cryptography]
bcrypt
python-dotenv
apscheduler
email-validator
